Apply Kalman filter

+++ include/pwm.h
@@ -12,7 +12,7 @@
 #include "nuttx/timers/pwm.h"
 
 void InitMotor(struct pwm_info_s *info);
-void PWMWriteAll(struct pwm_info_s *info, uint8_t perc1, uint8_t perc2, uint8_t perc3, uint8_t perc4);
+void PWMWriteAll(struct pwm_info_s *info, uint16_t perc1, uint16_t perc2, uint16_t perc3, uint16_t perc4);
 void PWMWriteSingle(struct pwm_info_s *info, int index, uint8_t percentile);
 
 #endif
